# Moving to TypeScript
I can't really take it any more, I just want good typing now.

## Moved CRA to typescript

Added recommended dependencies to package.json

Other changes:

- Updated index.js -> index.tsx
- starting small, updated transaction.js to ts
- currently tests dont pass due to no arg in new Muny()
- Muny.ts is next
