+++ package.json
@@ -4,15 +4,21 @@
   "private": true,
   "main": "electron-starter.js",
   "dependencies": {
+    "@types/jest": "^24.0.5",
+    "@types/node": "^11.9.4",
+    "@types/react": "^16.8.3",
+    "@types/react-dom": "^16.8.2",
     "bootstrap": "^4.2.1",
     "flux": "^3.1.3",
     "jquery": "^3.3.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-grid-layout": "^0.16.6",
-    "react-scripts": "2.1.3",
+    "react-scripts": "^2.1.5",
     "react-virtualized": "^9.21.0",
-    "reactstrap": "^7.1.0"
+    "reactstrap": "^7.1.0",
+    "redux": "^4.0.1",
+    "typescript": "^3.3.3"
   },
   "scripts": {
     "react:start": "react-scripts start",
